pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "requests>=2.32.3",
    "pymongo==4.4.0",
    "flask-pymongo==2.3.0",
    "trafilatura>=2.0.0",
    "bson>=0.5.10",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "firebase-admin>=6.8.0",
    "pyrebase4>=4.8.0",
]
